feat: implement repositories in CountryCoefficientService

+++ Core/Interfaces/ICountryCoefficientService.cs
@@ -2,7 +2,7 @@
 
 public interface ICountryCoefficientService
 {
-    Task<double> GetCountryPoints(string country);
+    Task<int> GetCountryPoints(string country);
     Task<bool> UpdateCountryCoefficient(string country, int countryNewPoints);
     Task<bool> RecalculateForMatchAsync(int matchId);
 }
